tests/wpt/web-platform-tests/navigation-timing/nav2_test_unloadEvents_no_previous_document.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>Navigation Timing 2 WPT</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
        <link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-navigation-timing-interface"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <script>

            function onload_test()
            {
                var frame = document.getElementById("frameContext");

                assert_equals(frame.contentWindow.performance.getEntriesByType("navigation")[0].type,
                    "navigate",
                    "Expected navigation type to be navigate.");

                assert_equals(frame.contentWindow.performance.getEntriesByType("navigation")[0].unloadEventStart, 0,
                    "Expected unloadEventStart to be 0.");
                assert_equals(frame.contentWindow.performance.getEntriesByType("navigation")[0].unloadEventEnd, 0,
                    "Expected unloadEventEnd to be 0.");
                done();
            }
        </script>

    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates the unload event times are 0 when there is no previous document.</p>

        <iframe id="frameContext" onload="onload_test();" src="resources/blank_page_green.html" style="width: 250px; height: 250px;"></iframe>
    </body>
</html>
